bb: support copy\paste

+++ platform/bb/rhodes/src/rhomobile/RhodesApplication.java
@@ -99,6 +99,7 @@ final public class RhodesApplication extends RhodesApplicationPlatform implement
 	public static final String LABEL_CLOSE = "Close";
 	public static final String LABEL_EXIT = "Exit";
 	public static final String LABEL_NONE = "none";
+	public static final String LABEL_COPYPASTE = "copy_paste";
 	
 	private static final RhoLogger LOG = RhoLogger.RHO_STRIP_LOG ? new RhoEmptyLogger() : 
 		new RhoLogger("RhodesApplication");
@@ -839,7 +840,8 @@ final public class RhodesApplication extends RhodesApplicationPlatform implement
     	}
     	
 		private Vector menuItems = new Vector();
-
+		private boolean m_bMenuCopyPaste = false;
+		
 		private MenuItem homeItem = new MenuItem("", 200000, 10) {
 			public void run() {
 					navigateHome();
@@ -895,8 +897,11 @@ final public class RhodesApplication extends RhodesApplicationPlatform implement
 				super.makeMenu(menu, instance);
 				return;
 			}
+			//return;
 			
-			menu.deleteAll();
+			if (!m_bMenuCopyPaste)
+				menu.deleteAll();
+			
 			// Don't draw menu if menuItems is null
 			if (menuItems == null)
 				return;
@@ -955,6 +960,8 @@ final public class RhodesApplication extends RhodesApplicationPlatform implement
     	    	setDefaultItemToMenuItems(label, closeItem);
     	    } else if (label.equalsIgnoreCase(RhodesApplication.LABEL_NONE)) {
     	    	menuItems = null;
+    	    } else if (value.equalsIgnoreCase(RhodesApplication.LABEL_COPYPASTE)) {
+    	    	m_bMenuCopyPaste = true;
     	    } else {
 				MenuItem itemToAdd = new MenuItem(label, 200000, 10) 
 				{
@@ -1029,6 +1036,7 @@ final public class RhodesApplication extends RhodesApplicationPlatform implement
 
 		public void setMenuItems(Vector menuItems) {
 			this.menuItems = menuItems;
+	    	m_bMenuCopyPaste = false;
 		}
 
 		public MenuItem getSavedGetLinkItem() {
